requirements.txt
####Python<3.11
streamlit==1.36.0
watchdog==4.0.1
joblib==1.3.2
##numpy==1.24.3
##pandas==1.5.3
pillow==10.2.0
##torch==2.1.2
#-----------------
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
fonttools==4.43.1
idna==3.4
importlib-resources==6.1.0
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.1
# ref: https://docs.streamlit.io/knowledge-base/dependencies/libgl
opencv-python==4.8.1.78
##opencv-contrib-python-headless
packaging==23.2
psutil==5.9.6
py-cpuinfo==9.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
pyyaml==6.0.1
requests==2.31.0
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
thop==0.1.1-2209072238
torch==1.13.1
##torch==1.9.0
torchvision==0.14.1
tqdm==4.66.1
typing-extensions==4.8.0
ultralytics==8.1.19
##ultralytics==8.0.199
urllib3==2.0.6
zipp==3.17.0